pyproject.toml
[tool.poetry]
name = "conf-finder"
version = "0.0.2"
description = "Configuration file finder."
authors = ["rcmdnk <rcmdnk@gmail.com>"]
repository = "https://github.com/rcmdnk/conf-finder"
homepage = "https://github.com/rcmdnk/conf-finder"
readme = "README.md"
license = "apache-2.0"
keywords = ["XGB", "XGB_CONFIG_HOME", "Git", "CONFIG"]
classifiers = []

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
gitpython = "^3.1.37"

[tool.poetry.group.dev.dependencies]
tomli = { version = "2.0.1", python = "<3.11"}
pytest = "7.4.2"
pytest-cov = "4.1.0"
pytest-xdist = "3.3.1"
pytest-benchmark = "4.0.0"
pyproject-pre-commit = "0.0.26"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-n auto"
testpaths = ["tests",]

[tool.black]
line-length = 79

[tool.autoflake]
remove-all-unused-imports = true
expand-star-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.autopep8]
ignore = "E203,E501,W503"
recursive = true
aggressive = 3

[tool.isort]
profile = "black"
line_length = 79

[tool.flake8]
# E203 is not PEP8 compliant and black insert space around slice: [Frequently Asked Questions - Black 22.12.0 documentation](https://black.readthedocs.io/en/stable/faq.html#why-are-flake8-s-e203-and-w503-violated)
# E501: Line too long. Disable it to allow long lines of comments and print lines which black allows.
# W503 is the counter part of W504, which follows current PEP8: [Line break occurred before a binary operator (W503)](https://www.flake8rules.com/rules/W503.html)
# D100~D106: Missing docstrings other than class (D101)
# D401: First line should be in imperative mood
ignore = "E203,E501,W503,D100,D102,D103,D104,D105,D106,D401"
max-complexity = 10
docstring-convention = "numpy"

[tool.bandit]
exclude_dirs = ["tests"]

[tool.mypy]
files = ["src/**/*.py"]
strict = true
warn_return_any = false
ignore_missing_imports = true
scripts_are_modules = true
install_types = true
non_interactive = true
implicit_reexport = true
